Crash demo part 2: Crash place indication
This shows the location of the crash, which is in turn a symptom of VM bug.
During marking an already swept pair is tried to be marked as live, but
the tag is incorrect #xa0 while the length of the array is #x52 so this
would cause an out-of-bounds access and potentially a segmentation
fault.

$ chibi-scheme -e '(include "crash-test.scm")'
crash test
----------
1. running VM for few million cycles for debugging:
    - started
    - done
2. filling stack to 2000 slots, adding launcher and filling heap:
    - started
    - done
3. pushing a triggering object to stack:
    - started
    - done
4. triggering one GC and releasing stack :
    - started
    - done
5. prefilling heap:
    - started
    - done
6. triggering the bug:
** CRITICAL ERROR: attempting to mark a pointer with an incorrect tag = a0, pointer = 204880 **
length of 'types' array = 52