Suggested fix version 2
This is the version that might be the same as the version 1. I could
imagine a scenario that the opcode itself is the last live vesion of
itself, so if it is not in stack it could be modified during sweeping
and that could have an issue during compilation of the opcode.